Don't send an SP_FILE message for pcapng unless we've written an SHB.
In the "pcapng passthrough" mode, we don't write an SHB to the capture
file as soon as we've opened it, as we're passing through the SHBs (and
IDBs) that we read from the pipe from which we're capturing.

If we're in "capture child" mode, our parent will try to open the
capture file as soon as it receives an SP_FILE message; that will end up
doing something such as treating the file as an empty ERF file when it
tries to open the empty file.

This means that we must not send the SP_FILE message until we've written
an SHB, if we're writing a pcapng file.

Make it so.

This should fix issue #17013.
(While we're at it, handle getting old-style deprecated Packet Blocks
from a pipe; we shouldn't get them, but we also should handle them if we
do get them.)