Bluetooth: Add connection parameters in frame context information to display
Add connection parameters in context information to display.
The connection parameters for the connection is important enough information
due to the time-division multiple access nature of Bluetooth, so having an in
frame pointer available is useful for debugging performance and timing related
problems.
Since the connection parameters are applied using instant then the event
counter is needed from the capture context in order to properly set this
information.
When this information is not available the connection parameters in frame is
set as active from the connection parameter update PDU, which is incorrect,
until the instant, but a good enough compromise without this information.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>