SOME/IP: Respect endianness parameter for strings
Actually do something with the endianness parameter for a string.
Based on the issue filed, assume that a string that is configured
as UTF-16LE also has a little endian length parameter if that length
parameter is more than one octet. (The spec is a little vague on this,
perhaps they could even be different endianness and there needs be
multiple configuration parameters?) Fixes #17058