Bluetooth BR/EDR RF: add FHS and LMP dissectors, reassemble and forward L2CAP
Prevoiusly, the BR/EDR RF dissector just decoded the packet header and
displayed the rest of the packet as blob. It made no effort to further
decode the contents.

This commit adds a dissector for the Bluetooth BR/EDR FHS packets
used to exchange frequency hopping parameters during link
setup. Contents from the FHS packets are used to populate a table of
devices, containing the full device addresses, and a connection table
that links Master LAP and LT_ADDRs to connections. Normally, only the
lower 24 bits of the device address (LAP, lower address part) are used
in the packet preamble, but the upper part (UAP) is needed for example
as initialization value of the CRC.

This commit adds a payload header decoder, to determine how to route
the contents to downstream dissectors.

This commit further adds a dissector for the Bluetooth BR/EDR
LMP protocol, which is used to manage the BR/EDR link(s). Data from
LMP is used to find out about secondary LT_ADDRs for a device for eSCO
connections, keeping track of eSCO packet sizes.

This commit further adds reassembly of fragmented L2CAP packets, and
routes L2CAP traffic to the existing L2CAP dissector.