dumpcap: stop sending EOF to tshark when using ring-buffer
In a previous commit a while back, the code that handles ring-buffer
file changes was moved to a function called
capture_loop_wrote_one_packet. This function is called by the
appropriate pcap_handler function. Originally this was handled
in the capture loop, capture_loop_start. For some reason, moving
this into the pcap handler has caused dumpcap to send an EOF
to tshark, causing tshark to stop when writing to STDOUT. Moving
this code back into the loop seems to fix the issue. I haven't
been able to thoroughly step through the code to see why dumpcap
sends the EOF.