Protobuf: Provide an option to add subdissectore for fields of all types
Currently Protobuf dissector allows to add subdissector for fields of byte/string type.
This change provide a checkbox to enable an option that allow to add a subdissector for all Protobuf fields.
This will be helpful if someone want to show more information about a filed in wireshark.
for example: a protobuf can have a int32 filed which can have a finite number of valid values with some application specific meaning.
A Lua dissector can show meaning of all the values of this filed with this option enabled.