eth: Make all MAC addresses show universal subfields (U/L, I/G, etc.)
Make all FT_ETHER fields have universal subfields for the U/L, I/G
bits, the OUI, and the overall address (resolved and not) the way
that Ethernet source and destination addresses do, so that one can
filter for a MAC address appearing in any protocol (similar to e212).

Handle the case where the address does not appear directly in the
tvb but is calculated (such as DVB-DATA MPE), and use
TRY_TO_FAKE_THIS_ITEM_OR_FREE() to add the subfields even when
the item is faked so that filtering works properly.

This does not add specific subfields of "{field name}.lg" and
"{field name}.ig", only universal ones, so it is related to issue #15493
and issue #17246 but perhaps does not close them.