Initial work on having packet options in wtap_rec
This work is for #14329, and is the first few steps in supporting
multiple comments per packet. Currently supports reading and writing
pcapng files that have multiple comments, but not
editing/adding/removing comments.

There was a comment in wiretap/wtap.h beside the lone pkt_comment member
that "this should become a full set of options." I took that to heart
and created wsutil/wstlv.[ch] . This creates a data type for a
type+length+value, and one for a list of same that can be passed around.
`wtap_rec->opt_comment` has been changed to `wtap_rec->options_list` to
hold the list of any and all options that the capture file loader adds
to each packet.

This is a first pass that compiles and "works." But:

* As noted, not all features are complete.
    * Lua support doesn't work yet.
    * There's no UI for editing, adding, or removing comments using this
      new model.
* The API could probably use refining and expansion.
* I've added lots of XXX comments for things that I need to address.
* Some sort of reference counting on data items would be a nice bonus.