fix #17369 tshark json now handles mixed children
tshark's JSON export used to assume that all nodes from a certain parent
have the same 'shape' - they all have children, or none have children.
This does not necessarily hold when the --no-duplicate-keys option is
activated: in that case, nodes with the same field name will be merged,
and we end up sometimes with fields containing arrays of mixed element
types (some having children, some not). When that happened, the output
used to be coerced to the element type of the first element of the list.
Now we check the element type for each element of the list, ensuring a
valid output.