ErlDP: add support for fragmented distribution messages
OTP-22 introduced support for fragmented distribution messages.
Add defragmentation and decoding support those.

The `flag` used in the distributed header collides with the flag
for FLOAT_EXT in the extermal term format. This makes is neccesary
to split the decoding of the distribution header tag from the
not Erlang terms.