ieee80211_radio: Use calculated bitrate if not supplied
For 802.11n if the bitrate is not supplied then the calculated
bitrate is used. This change does the same for 11ac and 11ax.

Sniffer traces taken on recent versions of Macos no longer supply
the bitrate for 11ac frames in the RADIOTAP header, this change
allows the wireless timeline to work with these traces.

Change-Id: Ibdd583f6a6a65403b206c941eb7fa31e7df798b0