can: Make dissector table can.subdissector (CAN ID) accessible for lua scripts
Change the way the decode as dissector table can.subdissector is created.
Creating the dissector table as FT_UINT32 makes it usable in lua scripts.

Before it was created by using register_decode_as_next_proto which creates a
dissector table of type FT_NONE which is not accessible in lua scripts.