Profile: Add Meta Information to profiles
If a file called "profile.json" is present in the profile directory, the information
for "author" and "version" is being displayed directly in the profile screen. If no
information is present, the fields are left empty.

Additionally, a field called "email" may be added, which will be displayed as a tooltip
for the author information