stream: conversation-based reassembly
Add reassembly functions that take a conversation_t as the data
parameter and use it, along with direction and a fragment ID, for
the reassembly hash but do not use addresses or ports from packet_info.

Use these functions with the stream API, where they fit the existing
stream keys. Make them available in the reassembly API for use with
other protocols which are endpoint based, particularly those containing
subpackets which can overwrite the addresses in pinfo (e.g. DVB with GSE
and MPE).