Proto: No unpriming of fields
This change enables fields that are primed manually via
proto_tree_prime_with_hfid() (without the detour via display filters) not to be
unprimed forcefully after their first occurance. This is helpful if you want to
prime a large number of fields without having to generate huge strings for a
display filter.

Relates to: #17264