UI: view associated waveform file
The Ellisys Vanguard Bluetooth traffic analyzer can display signal
waveforms in addition to the packet trace. It would be nice if
wirekshark offered a similar feature.

This patch implements a similar feature. If a waveform file with the
same base name and a waveform extension (such as .vcd or .fst), _and_
the pcapng file contains a special option that states the absolute
start time of the associated wave file, then wireshark offers the
option (in the view menu) to open the associated waveform by starting
the gtkwave waveform viewer. Once open, selecting a different packet
in wireshark automatically places the gtkwave marker at the time of
the packet, and approximately centers the waveform around the
marker. This is unidirectional, changing the waveform marker currently
does not update the wireshark packet selection.