wsdb: Export as wsdb (wireshark database) format
Packet details can be exported as wsdb (wireshark database).
A wsdb file a sqlite based database which stores the entire packet data and the
complete dissection tree including reassembled buffers. It is the intention to
have a file format to persist dissection results while having the data in a
structured way that they can be accessed quickly and easily even when for giant
files. (sqlite database tools or libraries)

To optimize export performance it implemented with a double buffered queue
(dissection thread and n commit threads) and can ba parallelized to use all
available cores. In case of a parallelized export the output is splitted into n
files while n is the number of cores -1. These files can be aggregated easily
in different ways (e.g. Sqlite command ATTACH or opening multiple database
connections).

To improve export performance even further the sqlite journal and disk
synchronization is switched off.

To optimize file size unique strings are stored in dedicated table to prevent
duplicates.

Field values are encoded as an integer, a double or string value. They may have
an additional representation value which is typically a calculated string to
display in the GUI.

Export can be accessed via GUI (File -> Export Packet Dissection).

A GUI tool called wsdb_browser can be used to browse wsdb files. It is on basic
prototype level right now but it is able the aggregate splitted wsdb files.