Proto: Enable fields to be forcefully marked as referenced
Introduce the possibility to mark fields as referenced by an API without the
need of a detour via a filter.

A field can have up to 5 different states.
A field may not be referenced at all.
On the other hand it can be referenced directly or indirectly. (Majorly driven
by filters, columns and so on) These references may be constructed and/or
removed frequently.
In addition to that a field can have an enforced direct or indirect reference.
With this it is possible it is possible to enforce a reference to a field
independently from the rest of the wireshark logic. An enforced reference needs
to be removed explicitly.

It is the intention to give taps the control which fields they want to get
provided by the protocol tree.