capture file: Better performance through caching of filter results
capture file implements a cache for filter results so that they can be accessed
further times way faster. In case of a cache match a rescan of the file is not
necessary. The cached filter result can be displayed to the user with almost no
noticable delay.

If a filter result is restored from the cache instead of rescanning the file
the user gets a temporary message shown in the statusbar.

The keys for the cache are strings which are primaly intended to be filter
strings.

GUI supports a preference to disable the cache and to limit the maximum number
of entries. By default the cache is disabled for now.

MR is related to #17547