IO Graph: Add checkbox to enable or disable automatic updates
Adds a checkbox 'Automatic Update' to the IO Graph to enable or disable
rescans and recalculation of graph data temporarily. This is usefull when
you want to modify settings of multiple graphs without triggering a rescan
with every change of a signale setting. This becomes usefull for large trace
files in particular because it can save a lot of waiting time.

Rescan or recalculation events are queued while 'Automatic Update' is not
active Unchecking 'Automatic Update' triggers the queued updates.