SMPP: Handle unaligned PDUs better
Test to see if we look unaligned and if so, consume all the data in
the current segment and mark it as continuation data.

When the heuristic dissector has found SMPP, mark it as a
conversation so that later segments that don't begin with a PDU
start are handled instead of being rejected.

There's room for more improvement by scanning through the current
segment to look for the next PDU start, but this makes it work
considerably better, at least as well as 1.10.x. Improves #11306.