WIP: reload file if loaded by lua file handler
Prevent certain failure when reloading lua plugins when the capture
file was loaded by a custom lua file handler. When reloading a
capture file from a lua file handler, reload the file before
redissecting the packets. Otherwise the deregistered FileHandler
will throw an error when rereading the packets for dissection.

Add a new wtap_uses_lua() API function to check if the current file
in the wiretap session is using a lua file handler.