dfilter: Parse protocol fields at the semantic level
Instead of parsing any blob of characters that matches a registered
protocol field as such in the lexer do it at the semantic level
because the meaning of "unparsed data" is contextual and sometimes
the meaning of unparsed does not include "protocol field", such as
on the right side of a matches relation.

For example the expression "udp.payload matches tcp" fails with
"protocol is invalid on the right-side of comparison", where the
intention of allowing unparsed strings (strings not encased in
double quotes) is to interpret the blob of bytes { 't', 'c', 'p'}
as a string in a context where protocol fields don't make sense
(for the sake of flexibility presumably).

A fix was added in ac0b1d42f314b752d87c9d583d6bd66e8140f534 for
check_relation() but instead of trying to catch the parse mistake
in one or more places remove the field concept from the grammar
and use the appropriate heuristic at the semantic level.

Ping #12810.