dfilter: Add unit test for "exists" and fix it
The expression "frame" is failing with "frame is not a valid
protocol" and this is not caught in the dfilter tests.

  % dftest "frame"
  Filter: "frame"
  dftest: "frame" is neither a field nor a protocol name.

Add a new display filter test for "exists" (test the existance
of a protocol field).

Fix the code by making sure an "exists" filter tries to parse
an unparsed string as a protocol field.

  % dftest "frame"
  Filter: "frame"

  Constants:

  Instructions:
  00000 CHECK_EXISTS	frame
  00001 RETURN

  % dftest "foobar"
  Filter: "foobar"
  dftest: "foobar" is neither a field nor a protocol name.

Add checkDfilterSucceed() to the test suite to test a successful
parse using dftest.

Put the new test in a new test group because this simple test didn't
seem to fit in any of the other groups. It might need a better name
than 'group_field' in the future.