dfilter: We need to handle FVALUEs in the parse tree
During semantic checking we transform semantic types to
concrete types. With reference counting the later can
appear in the parse tree from a previous pass if a
node is duplicated, i.e. its reference count is > 1.

An example would be the filter "ip.proto < 6 < 17".

The parse tree is LT(ip.proto,6) AND LT(6,17).

During the semantic check for the LHS of the AND test 6
has the semantic type UNPARSED. During the semantic check
of the RHS of the AND 6 is the same object as the LHS and
has the concrete type FVALUE[UINT8], not UNPARSED.

We need to handle that. Treat an FVALUE semantic check as
an STRING if the type is a string type or as UNPARSED
otherwise.