dfilter: Remove macro proto tree cache
This replaces the current macro reference system with
a different implementation. Instead of a macro a reference
is a syntax element. Note that this only makes sense in
contexts where there is a selected frame, because the
reference is always compiled with regard to the selected
frame. In other situations, like tshark, using a display
filter with a reference, in the format ${some_field}, will
generate an error. This is correct.

Because the new field reference syntac uses the same notation
as a macro we have to use a heuristic to distinguish them:
if the name has a dot it is a field reference, otherwise
it is a macro name.

Creating a macro cache of the whole proto tree is inefficient.
This is done even if no display filter is being used. Rewrite
the code to perform the reference substitution during the
compilation phase. Besides not creating a cache for every selection,
whether it uses a reference or not, references are substituted for
fvalues, not string fvalue representations.

The added complexity is that me must pass the dissections context
to dfilter_compile to resolve references at compilation time. And
we must ignore missing reference failures if we're just validating
the syntax.

This also improves the error message and adds strong type checking
to references (this is only visible in the GUI).

Before:

  $ tshark  -Y 'frame.number == ${frame.number}'
  tshark: macro 'frame.number' does not exist

After:

  $ tshark  -Y 'frame.number == ${frame.number}'
  tshark: No tree context to replace field reference frame.number.

The last point about strong type checking is because previously
references were compiled from strings (with the usual type coercions)
and now they are compiled from fields.