extcap: Improved debug log handling
Issue:
When --debug or --debug --debug-file is used, no logging is produced.
It is required to use --log-level=debug in addition to it, but when
used, output is written to requested file and to console (STDERR) too.
Output to console confuses wireshark/tshark then.

Changes:
When --debug used, log level is set to debug automatically.
If --debug-file used, log to console is disabled (discarded).
Change involves any application based on extcap/extcap-base.c.