Extcap/ssh-base: Use password authentication before keys
expcap/ssh-base.c uses common sequence to login to a device - keys and
then password authentication.
On many cisco devices, when ssh keys are tried, ssh connection fails
for no clear reason and libssh do not detect it. Therefore provided
password will fail.
I changed sequence to use password before keys.

The idea for change is simple. If user provided password, they are
sure it is necessary for successful login to the device.