Prefs/Extcap: Added support for password which is never stored on disk
Issue:
When a user sets extcap password, it is not remembered during runtime.
Therefore when the user runs extcap twice in row, password must be
entered over and over.

Solution:
New prefs type PREF_PASSWORD defined and extcap framework is adapted to
it.
It has welcomed consequence. Because now prefs key exists in prefs,
e.g. tshark can change it with -o option so it can run extcaps which
requires password. It was not possible to run such extcap with tshark
before.

Changes:
- PREF_PASSWORD handled nearly same way as PREF_STRING
  - new API calls added to register/set/get password
  - prefs key and description is stored in prefs file
  - password value is never stored nor read
  - UI editor shows password just during entering new password
    (PasswordEchoOnEdit), dots shown elsewhere
- extcap framework adapted to map EXTCAP_ARG_PASSWORD to PREF_PASSWORD
- updated tools/checkAPIs.pl to allow to ignore some cases in prefs_register_*_preference which are caused by this patch
- updated doc