Extcap prefs: Editor remembers empty values
Issue:
When any extcap parameter has nonempty default value, it is shown in
editor. If user clears the value (default value was useless for them),
next extcap run uses empty value for the parameter. When editor is
opened afterwards, empty value is again replaced by default value.
Therefore when user runs extcap multiple times in row, it is first time
started with empty value and later with default value.
Reason is that editor uses empty value as reset to default.

Solution:
Empty value is used as any other value and no reset to default is made.
But as consequence, there was no way how to reset do default value. So
editor is extended with "reset to default value" action next to every item
which is capable of it.

Changes:
- added reset to default value button for most of parameter types
- removed reseting of value to default when value is empty
- Restore Defaults button action adapted