WikiShark? Add ability to display protocol field values
In most cases Wireskark can be somehow like Wikipedia for protocols.
Because it shows protocols fields details. Add another level and show current
value and all other possible for specific field if available.
Feature for Qt Shark.

For example: ip.proto, for example 0x11 is UDP, but which is TCP? Now you can
read it in Wireshark GUI.

Change-Id: I6b58eb34be3bff603765e37c31fd8f929ec551ea