wsutil/nstime: Add parameter to parse ISO8601 in UTC
There are three possibe cases for the input format:
1. A date/time without timezone
2. A date/time with 'Z' timezone
3. A date/time with plus/minus timezone offset.

In the third case the function iso8601_to_nstime() always
returns a Unix time of UTC plus offset (local time?).
This is counter-intuitive. It would be more natural
to return UTC.

Instead of always returning time in UTC+offset (assumed
to be localtime), add a 'use_localtime' boolean option to
configure the return.

Optionally also return the timezone offset in seconds from UTC.