Differentiate `-c` from `-a packets:`
In TShark's man page, the description for `-a packets:N` implies that
the output file will *contain* N packets, but `-c N` implies only that
tshark will *read* N packets (which further implies the output file
could have fewer packets if `-Y` is used). However, the text also
falsely says the two are equivalent.

This patch causes `-a packets:N` to stop after the output *contains* N
packets when reading from one file to write to another. Fixes #17411.

Some changes were made to dumpcap for clarity with the new variable
names, but the actual behaviour of dumpcap is unchanged; due to the way
it's written, it's significantly more difficult to differentiate packets
seen from packets written.