16432: Specify directory for temporary captures
Add a capture option `--temp-dir` to wireshark/tshark/dumpcap that
specifies the directory in which temporary capture files are to be
written. If not specified, uses the system default behaviour.

Add a GUI element in the capture options dialog for specifying this
directory.

I found that this couldn't be effective as a user preference; it had to
be a capture option in order to be successfully propagated to dumpcap
child processes.