wsutil: New API to gather compile/runtime info
This new API is used in `ui/version_info.c` and by any place that calls
`ws_init_version_info()`. Current behaviour is to force the list of
present/absent dependencies into a long, unwieldy string; with this
change, dependencies are kept as a list until the time comes to format
them for output.

Lays the ground work to facilitate improving the layout of the "About"
information as mentioned at
<https://gitlab.com/wireshark/wireshark/-/issues/16270#note_840035739>