Proposed reworking of 'About:' info
Make the information in `tshark -v` output and Wireshark's "About"
dialog easier to read, by presenting present/absent dependencies in
columns (similar to UNIX `ls`) instead of as a run-on sentence.

The `+`/`-` prefix is kept to allow the test harness to be able to tell
whether a feature is present or not. Alternately, the test harness could
be made to look for features between a 'With:' and a (possibly not
present) 'Without:' in the output to determine their presence.