dfilter: Replace register preload with direct contant addressing.
Adds a new address space for constant values and dedicated
opcodes to use it.

This avoids the complicated logic to preload registers and manage
register lifecycles.

The MEMADDR DFVM value type is not strictly required because we
are using new opcodes but it is a nice explicit check and
avoids renaming the REGISTER value to something more generic.

This also makes the instructions to load constants into registers
obsolete, since they are addressed directly.

Rename ANY_IN_RANGE opcode because it is extremely confusing to have two different
kinds of ranges with the same name (slices and sets).