epan: Ressurect some really old, dead conversation code.
As #8010 points out, we have dead code whereever we have

    if (!(conversation->options & NO_ADDR_B) ...

or

    if (!(conversation->options & NO_PORT_B) ...

since at those points "conversation" points to entries in one of the
"no_addr2" or "no_port2" tables.

The suggested fix is to check the search options instead of the
conversation options. We originally did that, but commit 6bdffa496e
switched to conversation options.

Apply the fix from #8010 by hand, which revives the dead code. Use
non-negative logic where we can.

Fixes #8010.

Co-authored-by: Anton Sanarov <intkecsk@yandex.ru>