Qt: File Path Editor
The current situation in UatDelegate as well as PathChooserDelegate
leads to issues, where Wireshark crashes if the button is clicked. This
is due to the UI not correctly positioning the button inside the cell.

This change implements a widget, which will serve as cell content,
handling all relations with choosing the file, but also properly
handling the size and geometry of said cell content, therefore no longer
leading to crashes and cleaning up code at the same time, as duplicate
methods are being removed.

Fixes #17789
Fixes #17819
Fixes #18088