Lua: Improved API for obtaining fields
This extends existing all_field_infos() API by a parameter to specify
the wanted fields by their id or their filter name.

Secondly it introduces a new API Field.get_id() to obtain the field id
of a field given its filter name.

This can result in better performance as the user can request field
infos in a more specific way and conversions of unnecessary  fields are
avoided.

This becomes useful when you expect the same field multiple times in a
frame.