iec104: dissect the control field as a 4-octet little-endian field.
Treat all 4 octets of the control field as a single little-endian value
divided into bitfields.  We already showed *some* subfields as
bitfields; this means we show *all* of them that way.

That makes the display more clearly show which bits in those octets
correspond to which fields.

It also fixes the dissection of the type field; we have separate
bitfields for I frames (1-bit bitfield) and S and U frames (2-bit
bitfield).

Use proto_tree_add_item_ret_uint() to fetch the values other than the
frame type value.

Fixes #18167.