tools: Remove dead/unused code from make-version.pl.
Remove the "ugly internal SVN file hack" block from make-version.pl. It
only worked for Subversion versions prior to 1.4, which was released in
2006.

Remove the Bazaar (bzr) block. James Brown may not have sang about it,
but even so it's a Git's world.

Remove the Tortoise block. As noted in a comment, it appears to be dead
code.