Add field for DIAMETER command code text [#16833]
Add a string field to the DIAMETER dissector which provides the string
value of the command code. This makes it possible to search for, add as
a column, etc.

The `value_string`s used to generate these values are determined at
run-time, so they cannot easily be specified for the `hf_diameter_code`
field. Adding it as a separate field is much easier.