Handle invalid bytes passed to print_escaped_xml function
Some Wireshark decoders provide byte sequences to the
print_escaped_xml() function that cannot be handled with the existing
conversions implemented in this function. According to the XML spec -
https://www.xml.com/axml/target.html#charsets - XML text is a sequence
of valid characters, and a valid character may be a tab, carriage
return, line feed, or any of the legal graphic characters of Unicode:

Char ::= #x9 | #xA | #xD | [#x20-#xD7FF] |
         [#xE000-#xFFFD] |[#x10000-#x10FFFF]

Some decoders provide byte sequences containing byte values less than
int 31, but that are not any of tab, CR or LF. In that case, the current
implementation of print_escaped_xml() sends them through to the encoded
result unchanged. This results in invalid XML. For an example, try the
following:

$ wget https://storage.googleapis.com/gcla3/xmlbug.pcapng
$ tshark -r xmlbug.pcapng -T pdml | xmllint --noout - || echo bad xml

This PR modifies the function so that these invalid bytes are converted
by "escaping" them e.g. the byte with int value 28 decimal is serialized
as "\01c" (without the quotes). But other choices may be reasonable and
better than this one!