Diameter: Allow diameter command code text to be added to custom column
Create the command value string when loading the dictionary and
assign it to the command code field when registering it. This
makes it behave like any other value_string for adding to a custom
column (allowing resolved and unresolved), etc.

It could be changed to use a wmem_array instead of a GArray, as is
currently done with the AVPs.

Fix #16833. (Note that solving the analogous issue for the AVP
code text is more complicated, since the interpretation depends
on the vendor ID.)