ISUP tap with proper message
The current ISUP tap handler is hardcoded to generate messages from the
default ISUP `value_string_ext`, regardless of what variant is in use.
In addition, it doesn't handle messages that aren't in the
`value_string_ext`.

This change store the ISUP variant in the tap record, and uses it to
generate the appropriate message for the variant in use. In addition, it
follows behaviour elsewhere in the dissector and labels unknown messages
as "reserved".