IO Graph: Plot all field instances
Allows to plot all instances of a field in as time series via
"DIRECT(Y Field)".

No aggregation function like SUM, MAX, MIN or others is applied.
The value of the field are plotted directly.

Since all instances of a field are plotted it independant of selected
interval and SMA calculation.

It is limited to numeric fields.