Filter cache
Filter results can be cached. If a filter was used previously and the cache
has a match an entire rescan can prevented and instead the filtered packet list
can be restored from the cache. The cache is limited to be used with capture
files.

If a filter result is restored from the cache instead of rescanning the file
the user gets a temporary message shown in the statusbar.

The maximum size of the cache can be customized by the user via prefenrence.
The cache is can be disabled by by setting the maximum size to 0.
By default the cache is disabled.