Qt: Use deleteLater() to delete context menus.
A recent change set the WA_DeleteOnClose attribute on many of our
context menus. It looks like this can cause crashes in some situations,
for example when we try to use sender() in the code triggered by the
menu as suggested by the stack traces in #18282 and #18315.

Connect the menu's triggered signal to deleteLater instead of setting
WA_DeleteOnClose. This should avoid any race conditions we might run
into when trying to access context menus or their actions after they've
been triggered. Delete our context menus in a couple of places where we
weren't doing so.

Fixes #18251 and related bugs.